added the draggable component, drag target, and the ability of moving a project.
Everything working as should be.

Remember by default JavaScript prevents the ability to drop a draggable item into a drag target; so in the drag-over handler of that same drag target element, you must first invoke event.preventDefault, to remove the default behaviour i.e. not being able to drop an item.